requirements.txt
Pillow
google-genai
dotenv
fastapi
uvicorn